+++ Assets/Scripts/Camera/CameraManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5afd2ad0c8d6d1049b249a187f46e271
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
